requirements.txt
fastapi==0.109.2
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
uvicorn==0.27.1
bcrypt==4.1.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
httpx==0.26.0
pytest
python-decouple==3.8